Split file.ts
