+++ connectors/src/connectors/google_drive/temporal/file/index.ts
@@ -1,0 +1,1 @@
+export { syncOneFile } from "./sync_one_file";
